pyproject.toml
[tool.poetry]
name = "arca"
version = "0.1.0"
description = "A lightweight CMS using FastHTML and Markdown/YAML files"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "arca"}]

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.12.1"
pyyaml = "^6.0.1"
markdown = "^3.6.7"
python-frontmatter = "^1.0.0"
watchdog = "^2.2.1"
pydantic = "^2.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
arca = "arca.cli:main"